04-Multitasking: Fix processor mode in first activate
As mention in issue #27, `task_init()` is added before
task created.
This is reference to 06-Preemptive, and also, it's not
the time yet to disclose the systick interrupt so use
`svc` to switch to handler mode after reset.

In previous way, we call the `activate` in both thread
and handler mode, and use `if(first)` to distinguish them.
This always bother the newbie to mini-arm-os to figure out
what `activate` is going on.

This commit makes `activate` only effect in handler mode
as I use `task_init()` to ""init"" reset mode from thread mode
to handler mode.